# Speed up copy!(dest, Rdest, src, Rsrc) by splitting out first dimension