I noticed something also reported by @JKrehl: our "block copier" `copy!(dest, Rdest, src, Rsrc)` (where `Rdest` and `Rsrc` are CartesianRanges) is slow. Presumably #9080 has something to do with it.

Master:
```julia
julia> A = rand(200, 200, 16);

julia> B = similar(A);

julia> R = CartesianRange(indices(A))
CartesianRange{CartesianIndex{3}}(CartesianIndex{3}((1, 1, 1)), CartesianIndex{3}((200, 200, 16)))

julia> copy!(B, R, A, R);

julia> @benchmark copy!(B, R, A, R) seconds=1
BenchmarkTools.Trial: 
  memory estimate:  0 bytes
  allocs estimate:  0
  --------------
  minimum time:     3.379 ms (0.00% GC)
  median time:      3.499 ms (0.00% GC)
  mean time:        3.591 ms (0.00% GC)
  maximum time:     6.628 ms (0.00% GC)
  --------------
  samples:          273
  evals/sample:     1
  time tolerance:   5.00%
  memory tolerance: 1.00%
```
This PR:
```julia
julia> @benchmark copy!(B, R, A, R) seconds=1
BenchmarkTools.Trial: 
  memory estimate:  0 bytes
  allocs estimate:  0
  --------------
  minimum time:     688.652 μs (0.00% GC)
  median time:      838.804 μs (0.00% GC)
  mean time:        857.291 μs (0.00% GC)
  maximum time:     1.689 ms (0.00% GC)
  --------------
  samples:          1074
  evals/sample:     1
  time tolerance:   15.00%
  memory tolerance: 1.00%
```
So approximately 4x faster for a 3d array.

I also decided to write a docstring for this method, and to explicitly test it. (I guess previously it had been tested only indirectly.)