@JKrehl 2017-03-08
The bottleneck in my view is the parsing of the indices into the `arrayref` where, I suspect, the length tuple generated by iterating the `CartesianRange` is resolved at runtime. By unpacking the tuple and the iteration-loops explicitly (at compile time) the speedup is even more substantial as well as having less code.

```julia
@generated function copy!{T,N}(dest::AbstractArray{T,N}, Rdest::CartesianRange{CartesianIndex{N}}, src::AbstractArray{T,N}, Rsrc::CartesianRange{CartesianIndex{N}})
    quote
        isempty(Rdest) && return dest
        if size(Rdest) != size(Rsrc)
            throw(ArgumentError("source and destination must have same size (got $(size(Rsrc)) and $(size(Rdest)))"))
        end
        @boundscheck checkbounds(dest, Rdest.start)
        @boundscheck checkbounds(dest, Rdest.stop)
        @boundscheck checkbounds(src, Rsrc.start)
        @boundscheck checkbounds(src, Rsrc.stop)
        ΔI = Rdest.start - Rsrc.start
        @nloops $N i (n->Rsrc.start[n]-start(indices(Rsrc)[n])+indices(Rsrc)[n]) begin
            @inbounds @nref($N,dest,n->i_n+ΔI[n]) = @nref($N,src,i)
        end
        dest
    end
end
```
Wih the timings:
```julia
In[34]: precompile(copy!, map(typeof, (B, R, A, R)))
@benchmark copy!($B, $R, $A, $R) seconds=1
Out[34]: BenchmarkTools.Trial: 
  memory estimate:  0 bytes
  allocs estimate:  0
  --------------
  minimum time:     373.251 μs (0.00% GC)
  median time:      383.829 μs (0.00% GC)
  mean time:        402.241 μs (0.00% GC)
  maximum time:     730.292 μs (0.00% GC)
  --------------
  samples:          2370
  evals/sample:     1
  time tolerance:   5.00%
  memory tolerance: 1.00%
￼
In[48]: precompile(copy!, map(typeof, (B, R, A, R)))
@benchmark copy!($B, $R, $A, $R) seconds=1
Out[48]:
BenchmarkTools.Trial: 
  memory estimate:  0 bytes
  allocs estimate:  0
  --------------
  minimum time:     2.665 ms (0.00% GC)
  median time:      2.768 ms (0.00% GC)
  mean time:        2.824 ms (0.00% GC)
  maximum time:     3.380 ms (0.00% GC)
  --------------
  samples:          351
  evals/sample:     1
  time tolerance:   5.00%
  memory tolerance: 1.00%
```
There is no `indices(::CartesianRange, ::Int)` defined, is that issue-worthy?
Wouldn't it be sensible to define `indices(::CartesianRange)` as between `start` and 'stop' rather than 1 and 'size'?