@timholy 2017-03-08
Your version has the same performance as mine when I run yours on my machine. Are you comparing timings on your machine versus timings on my machine? Or are you genuinely faster on yours?

There's a push to get rid of `@generated` functions wherever possible, so I'm afraid that even if this is faster, depending on `Base.Cartesian` is probably not something we want to do here.