@KristofferC 2017-03-08
> There's a push to get rid of `@generated` functions wherever possible

While this is true, this seems like a perfect usage of non-dangerous `@generated` (pure loop generation). I haven't benchmarked myself so if the two methods are of the same speed then of course, go with the non `@generated`, but it feels a bit silly to leave performance on the table because of not wanting to use a language feature in the intended way because of "reasons".