@JKrehl 2017-03-08
If I correct for the difference in the `Base.copy!` (I forgot to re-rename it in my pasted timings), the difference is still a factor of 0.7.

Overall I do think my solution is much more compact and straight-forward as it tackles the problem of speed difference between `A[5,5,5]` and `A[(5,5,5)...]` (which is about factor 3.5 in my mini-test) directly. Your solution does that to but only for the first dimension.

That fundamental problem could be fixed by defining `getindex{T,N}(::AbstractArray{T,N}, ::NTuple{N,Int})` or something like that so the index tuple does not need to be parsed at access time. Such a solution could allow an equally fast solution without a `@generated` function (BTW I do think that is exactly what they are for if equivalent code cannot be generated by normal syntactic means).