@timholy 2017-03-08
It's parsed at compile time.
```julia
julia> A = rand(5,5,5);

julia> I = CartesianIndex(5,5,5);

julia> foo1(A, I) = (@inbounds ret = A[I]; ret)
foo1 (generic function with 1 method)

julia> foo3(A, i, j, k) = (@inbounds ret = A[i,j,k]; ret)
foo3 (generic function with 1 method)

julia> @code_native foo1(A, I)
        .text
Filename: REPL[3]
        pushq   %rbp
        movq    %rsp, %rbp
        movq    (%rdi), %rax
Source line: 1
        movq    8(%rsi), %rcx
        movq    16(%rsi), %rdx
        addq    $-1, %rdx
        imulq   32(%rdi), %rdx
        leaq    -1(%rcx,%rdx), %rcx
        imulq   24(%rdi), %rcx
        addq    (%rsi), %rcx
        vmovsd  -8(%rax,%rcx,8), %xmm0  # xmm0 = mem[0],zero
        popq    %rbp
        retq
        nopl    (%rax)

julia> @code_native foo3(A, 5, 5, 5)
        .text
Filename: REPL[4]
        pushq   %rbp
        movq    %rsp, %rbp
        movq    (%rdi), %rax
Source line: 1
        addq    $-1, %rcx
        imulq   32(%rdi), %rcx
        leaq    -1(%rdx,%rcx), %rcx
        imulq   24(%rdi), %rcx
        addq    %rsi, %rcx
        vmovsd  -8(%rax,%rcx,8), %xmm0  # xmm0 = mem[0],zero
        popq    %rbp
        retq
        nopw    %cs:(%rax,%rax)
```

Are you sure it's faster? Maybe build this branch to be certain?

That said, the Base.Cartesian solution does circumvent #9080 completely, not just for the first dimension. It would matter if the first dimension is tiny.