@timholy 2017-03-08
> There is no indices(::CartesianRange, ::Int) defined, is that issue-worthy? Wouldn't it be sensible to define indices(::CartesianRange) as between start and 'stop' rather than 1 and 'size'?

An oversight that's definitely issue-worthy (would be nicer than the `convert` version). Thanks for noticing! I can just add it to this, though, no need to file an issue.