@timholy 2017-03-08
One problem, though:
```julia
julia> A = rand(3,5);

julia> R = CartesianRange(indices(A));

julia> indices(A)
(Base.OneTo(3), Base.OneTo(5))

julia> indices(R)
(1:3, 1:5)
```
So we lose the type information. Consequence:
```julia
julia> similar(A, indices(A))
3×5 Array{Float64,2}:
 6.90335e-310  6.90335e-310  6.90335e-310  6.90335e-310  6.90334e-310
 6.90335e-310  6.90335e-310  6.90335e-310  6.90334e-310  6.90334e-310
 6.90335e-310  6.90335e-310  6.90335e-310  6.90334e-310  0.0         

julia> similar(A, indices(R))
ERROR: MethodError: no method matching similar(::Array{Float64,2}, ::Type{Float64}, ::Tuple{UnitRange{Int64},UnitRange{Int64}})
Closest candidates are:
  similar(::Array{T,2}, ::Type) where T at array.jl:179
  similar(::Array, ::Type, ::Tuple{Vararg{Int64,N}}) where N at array.jl:181
  similar(::AbstractArray, ::Type{T}) where T at abstractarray.jl:507
  ...
Stacktrace:
 [1] similar(::Array{Float64,2}, ::Tuple{UnitRange{Int64},UnitRange{Int64}}) at ./abstractarray.jl:508

julia> using OffsetArrays

julia> similar(A, indices(R))
OffsetArrays.OffsetArray{Float64,2,Array{Float64,2}} with indices 1:3×1:5:
 6.90335e-310  6.90335e-310  6.90335e-310  6.90335e-310  6.90335e-310
 6.90334e-310  6.90334e-310  6.90334e-310  6.90335e-310  6.90335e-310
 6.90335e-310  6.90334e-310  6.90335e-310  6.90335e-310  6.90335e-310
```

I think we should rewrite `CartesianRange` so it's just a tuple of `AbstractUnitRange` objects. But we can't do that for 0.6 at this point, since that would be a breaking change.